requirements.txt
python-telegram-bot==20.3
ccxt==4.2.5
python-dotenv==1.0.0
